requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
matplotlib
seaborn
numpy
csv

